+++ tests/unit/components/ModelizerTextView.spec.js
@@ -9,10 +9,12 @@ installQuasarPlugin();
 jest.mock('src/composables/Project', () => ({
   getProjectFiles: jest.fn(() => Promise.resolve([{ path: 'terraform/app.tf' }])),
   readProjectFile: jest.fn(() => Promise.resolve({ path: 'terraform/app.tf', content: 'new content' })),
-  getProjectName: jest.fn(() => 'projectName'),
 }));
 
 jest.mock('src/composables/FileExplorer', () => ({
+  createFile: jest.fn(() => 'createFile'),
+  createFolder: jest.fn(() => 'createFolder'),
+  sortTreeElements: jest.fn(() => 'sortTreeElements'),
   getTree: jest.fn(() => (
     [{
       id: 'projectName',
@@ -35,8 +37,21 @@ jest.mock('src/composables/FileExplorer', () => ({
 jest.mock('src/composables/events/FileEvent', () => ({
   OpenFileEvent: {
     subscribe: jest.fn(),
+    next: jest.fn(),
+  },
+  SelectNodeEvent: {
+    subscribe: jest.fn(),
+  },
+  CreateFileEvent: {
+    subscribe: jest.fn(),
+  },
+  DeleteFileEvent: {
+    subscribe: jest.fn(),
   },
   SelectFileEvent: {
+    next: jest.fn(),
+  },
+  ExpandFolderEvent: {
     next: jest.fn(),
   },
 }));
@@ -54,6 +69,12 @@ describe('Test component: ModelizerTextView', () => {
   let wrapper;
   let openFileSubscribe;
   let openFileUnsubscribe;
+  let selectNodeSubscribe;
+  let selectNodeUnsubscribe;
+  let createFileSubscribe;
+  let createFileUnsubscribe;
+  let deleteFileSubscribe;
+  let deleteFileUnsubscribe;
   let updateRemoteSubscribe;
   let updateRemoteUnsubscribe;
   let checkoutSubscribe;
@@ -63,6 +84,12 @@ describe('Test component: ModelizerTextView', () => {
   beforeEach(() => {
     openFileSubscribe = jest.fn();
     openFileUnsubscribe = jest.fn();
+    selectNodeSubscribe = jest.fn();
+    selectNodeUnsubscribe = jest.fn();
+    createFileSubscribe = jest.fn();
+    createFileUnsubscribe = jest.fn();
+    deleteFileSubscribe = jest.fn();
+    deleteFileUnsubscribe = jest.fn();
     updateRemoteSubscribe = jest.fn();
     updateRemoteUnsubscribe = jest.fn();
     checkoutSubscribe = jest.fn();
@@ -75,6 +102,18 @@ describe('Test component: ModelizerTextView', () => {
     FileEvent.OpenFileEvent.subscribe.mockImplementation(() => {
       openFileSubscribe();
       return { unsubscribe: openFileUnsubscribe };
+    });
+    FileEvent.SelectNodeEvent.subscribe.mockImplementation(() => {
+      selectNodeSubscribe();
+      return { unsubscribe: selectNodeUnsubscribe };
+    });
+    FileEvent.CreateFileEvent.subscribe.mockImplementation(() => {
+      createFileSubscribe();
+      return { unsubscribe: createFileUnsubscribe };
+    });
+    FileEvent.DeleteFileEvent.subscribe.mockImplementation(() => {
+      deleteFileSubscribe();
+      return { unsubscribe: deleteFileUnsubscribe };
     });
     GitEvent.CheckoutEvent.subscribe.mockImplementation(() => {
       checkoutSubscribe();
@@ -108,87 +147,87 @@ describe('Test component: ModelizerTextView', () => {
   });
 
   describe('Test function: onOpenFileEvent', () => {
-    it('should set a new value to activeFile', () => {
-      wrapper.vm.activeFile = { isSelected: false, id: '' };
+    it('should set a new value to activeFileTab', () => {
+      wrapper.vm.activeFileTab = { isSelected: false, id: '' };
       wrapper.vm.onOpenFileEvent({ id: 'terraform/app.tf', label: 'app.tf', content: '' });
 
-      expect(wrapper.vm.activeFile).toEqual({ isSelected: true, id: 'terraform/app.tf' });
+      expect(wrapper.vm.activeFileTab).toEqual({ isSelected: true, id: 'terraform/app.tf' });
     });
 
-    it('should push the new file to files Array if it is not already there', () => {
-      wrapper.vm.files = [{ id: 'terraform/app.tf', label: 'app.tf', content: '' }];
+    it('should push the new file to fileTabArray if it is not already there', () => {
+      wrapper.vm.fileTabArray = [{ id: 'terraform/app.tf', label: 'app.tf', content: '' }];
       wrapper.vm.onOpenFileEvent({ id: 'README.md', label: 'README.md', content: '' });
 
-      expect(wrapper.vm.files).toEqual([{ id: 'terraform/app.tf', label: 'app.tf', content: '' }, { id: 'README.md', label: 'README.md', content: '' }]);
+      expect(wrapper.vm.fileTabArray).toEqual([{ id: 'terraform/app.tf', label: 'app.tf', content: '' }, { id: 'README.md', label: 'README.md', content: '' }]);
     });
 
-    it('should not push the new file to files Array if it is already there', () => {
-      wrapper.vm.files = [{ id: 'terraform/app.tf', label: 'app.tf', content: '' }];
+    it('should not push the new file to fileTabArray if it is already there', () => {
+      wrapper.vm.fileTabArray = [{ id: 'terraform/app.tf', label: 'app.tf', content: '' }];
       wrapper.vm.onOpenFileEvent({ id: 'terraform/app.tf', label: 'app.tf', content: '' });
 
-      expect(wrapper.vm.files).toEqual([{ id: 'terraform/app.tf', label: 'app.tf', content: '' }]);
+      expect(wrapper.vm.fileTabArray).toEqual([{ id: 'terraform/app.tf', label: 'app.tf', content: '' }]);
     });
   });
 
-  describe('Test function: closeFile', () => {
+  describe('Test function: deleteFileTab', () => {
     beforeEach(() => {
       emit = jest.fn();
 
       FileEvent.SelectFileEvent.next.mockImplementation(() => emit());
     });
 
-    describe('The file to close is contained within files Array', () => {
-      it('should remove file if files length is 2 or more', () => {
-        wrapper.vm.files = [{ id: 'terraform/app.tf', label: 'app.tf', content: '' }, { id: 'README.md', label: 'README.md', content: '' }];
-        wrapper.vm.closeFile('terraform/app.tf');
+    describe('The file to close is contained within fileTabArray', () => {
+      it('should remove file if fileTabArray length is 2 or more', () => {
+        wrapper.vm.fileTabArray = [{ id: 'terraform/app.tf', label: 'app.tf', content: '' }, { id: 'README.md', label: 'README.md', content: '' }];
+        wrapper.vm.deleteFileTab('terraform/app.tf');
 
-        expect(wrapper.vm.files).toEqual([{ id: 'README.md', label: 'README.md', content: '' }]);
+        expect(wrapper.vm.fileTabArray).toEqual([{ id: 'README.md', label: 'README.md', content: '' }]);
       });
 
-      it('should remove file if files length is 1', () => {
-        wrapper.vm.files = [{ id: 'terraform/app.tf', label: 'app.tf', content: '' }];
-        wrapper.vm.closeFile('terraform/app.tf');
+      it('should remove file if fileTabArray length is 1', () => {
+        wrapper.vm.fileTabArray = [{ id: 'terraform/app.tf', label: 'app.tf', content: '' }];
+        wrapper.vm.deleteFileTab('terraform/app.tf');
 
-        expect(wrapper.vm.files).toEqual([]);
+        expect(wrapper.vm.fileTabArray).toEqual([]);
       });
     });
 
-    describe('Closing the active file, then files Array is empty', () => {
-      it('should reset and emit activeFile initial value', () => {
+    describe('Closing the active file, then fileTabArray is empty', () => {
+      it('should reset and emit activeFileTab initial value', () => {
         expect(emit).not.toHaveBeenCalled();
-        wrapper.vm.files = [{ id: 'README.md', label: 'README.md', content: '' }];
-        wrapper.vm.activeFile = { isSelected: true, id: 'README.md' };
-        wrapper.vm.closeFile('README.md');
+        wrapper.vm.fileTabArray = [{ id: 'README.md', label: 'README.md', content: '' }];
+        wrapper.vm.activeFileTab = { isSelected: true, id: 'README.md' };
+        wrapper.vm.deleteFileTab('README.md');
 
-        expect(wrapper.vm.files).toHaveLength(0);
-        expect(wrapper.vm.activeFile).toEqual({ isSelected: false, id: '' });
+        expect(wrapper.vm.fileTabArray).toHaveLength(0);
+        expect(wrapper.vm.activeFileTab).toEqual({ isSelected: false, id: '' });
         expect(emit).toHaveBeenCalledTimes(1);
       });
     });
 
-    describe('Closing the active file, then files Array is not empty', () => {
-      it('should set and emit activeFile new value', () => {
+    describe('Closing the active file, then fileTabArray is not empty', () => {
+      it('should set and emit activeFileTab new value', () => {
         expect(emit).not.toHaveBeenCalled();
-        wrapper.vm.files = [{ id: 'terraform/app.tf', label: 'app.tf', content: '' }, { id: 'README.md', label: 'README.md', content: '' }];
-        wrapper.vm.activeFile = { isSelected: true, id: 'README.md' };
-        wrapper.vm.closeFile('README.md');
+        wrapper.vm.fileTabArray = [{ id: 'terraform/app.tf', label: 'app.tf', content: '' }, { id: 'README.md', label: 'README.md', content: '' }];
+        wrapper.vm.activeFileTab = { isSelected: true, id: 'README.md' };
+        wrapper.vm.deleteFileTab('README.md');
 
-        expect(wrapper.vm.files).toEqual([{ id: 'terraform/app.tf', label: 'app.tf', content: '' }]);
-        expect(wrapper.vm.files).toHaveLength(1);
-        expect(wrapper.vm.activeFile).toEqual({ isSelected: true, id: 'terraform/app.tf' });
+        expect(wrapper.vm.fileTabArray).toEqual([{ id: 'terraform/app.tf', label: 'app.tf', content: '' }]);
+        expect(wrapper.vm.fileTabArray).toHaveLength(1);
+        expect(wrapper.vm.activeFileTab).toEqual({ isSelected: true, id: 'terraform/app.tf' });
         expect(emit).toHaveBeenCalledTimes(1);
       });
     });
 
     describe('Closing a non active file', () => {
-      it('should not set and not emit a new value to activeFile', () => {
-        wrapper.vm.files = [{ id: 'terraform/app.tf', label: 'app.tf', content: '' }, { id: 'README.md', label: 'README.md', content: '' }];
-        wrapper.vm.activeFile = { isSelected: true, id: 'terraform/app.tf' };
-        wrapper.vm.closeFile('README.md');
+      it('should not set and not emit a new value to activeFileTab', () => {
+        wrapper.vm.fileTabArray = [{ id: 'terraform/app.tf', label: 'app.tf', content: '' }, { id: 'README.md', label: 'README.md', content: '' }];
+        wrapper.vm.activeFileTab = { isSelected: true, id: 'terraform/app.tf' };
+        wrapper.vm.deleteFileTab('README.md');
 
-        expect(wrapper.vm.files).toEqual([{ id: 'terraform/app.tf', label: 'app.tf', content: '' }]);
-        expect(wrapper.vm.files).toHaveLength(1);
-        expect(wrapper.vm.activeFile).toEqual({ isSelected: true, id: 'terraform/app.tf' });
+        expect(wrapper.vm.fileTabArray).toEqual([{ id: 'terraform/app.tf', label: 'app.tf', content: '' }]);
+        expect(wrapper.vm.fileTabArray).toHaveLength(1);
+        expect(wrapper.vm.activeFileTab).toEqual({ isSelected: true, id: 'terraform/app.tf' });
         expect(emit).not.toHaveBeenCalled();
       });
     });
@@ -242,38 +281,70 @@ describe('Test component: ModelizerTextView', () => {
       ]);
     });
 
-    it('should update files', async () => {
-      wrapper.vm.files = [{ id: 'terraform/app.tf', label: 'app.tf', content: 'content' }];
+    it('should update fileTabArray', async () => {
+      wrapper.vm.fileTabArray = [{ id: 'terraform/app.tf', label: 'app.tf', content: 'content' }];
       wrapper.vm.updateProjectFiles();
       await flushPromises();
-      expect(wrapper.vm.files).toEqual([{ id: 'terraform/app.tf', content: 'new content', label: 'app.tf' }]);
+      expect(wrapper.vm.fileTabArray).toEqual([{ id: 'terraform/app.tf', content: 'new content', label: 'app.tf' }]);
     });
 
-    it('should update activeFile if it is no more contained in updated files', async () => {
-      wrapper.vm.files = [{ id: 'README.md', label: 'README.md', content: 'content' }, { id: 'terraform/app.tf', label: 'app.tf', content: 'content' }];
-      wrapper.vm.activeFile = { isSelected: true, id: 'README.md' };
+    it('should update activeFileTab if it is no more contained in updated fileTabArray', async () => {
+      wrapper.vm.fileTabArray = [{ id: 'README.md', label: 'README.md', content: 'content' }, { id: 'terraform/app.tf', label: 'app.tf', content: 'content' }];
+      wrapper.vm.activeFileTab = { isSelected: true, id: 'README.md' };
 
       wrapper.vm.updateProjectFiles();
       await flushPromises();
-      expect(wrapper.vm.activeFile).toEqual({ isSelected: true, id: 'terraform/app.tf' });
+      expect(wrapper.vm.activeFileTab).toEqual({ isSelected: true, id: 'terraform/app.tf' });
     });
 
-    it('should reset activeFile if updated files is empty', async () => {
-      wrapper.vm.files = [{ id: 'README.md', label: 'README.md', content: 'content' }];
-      wrapper.vm.activeFile = { isSelected: true, id: 'README.md' };
+    it('should reset activeFileTab if updated fileTabArray is empty', async () => {
+      wrapper.vm.fileTabArray = [{ id: 'README.md', label: 'README.md', content: 'content' }];
+      wrapper.vm.activeFileTab = { isSelected: true, id: 'README.md' };
 
       wrapper.vm.updateProjectFiles();
       await flushPromises();
-      expect(wrapper.vm.activeFile).toEqual({ isSelected: false, id: '' });
+      expect(wrapper.vm.activeFileTab).toEqual({ isSelected: false, id: '' });
     });
 
-    it('should not update activeFile if it is still contained in updated files', async () => {
-      wrapper.vm.files = [{ id: 'terraform/app.tf', label: 'app.tf', content: 'other content' }, { id: 'README.md', label: 'README.md', content: 'content' }];
-      wrapper.vm.activeFile = { isSelected: true, id: 'terraform/app.tf' };
+    it('should not update activeFileTab if it is still contained in updated fileTabArray', async () => {
+      wrapper.vm.fileTabArray = [{ id: 'terraform/app.tf', label: 'app.tf', content: 'other content' }, { id: 'README.md', label: 'README.md', content: 'content' }];
+      wrapper.vm.activeFileTab = { isSelected: true, id: 'terraform/app.tf' };
 
       wrapper.vm.updateProjectFiles();
       await flushPromises();
-      expect(wrapper.vm.activeFile).toEqual({ isSelected: true, id: 'terraform/app.tf' });
+      expect(wrapper.vm.activeFileTab).toEqual({ isSelected: true, id: 'terraform/app.tf' });
+    });
+  });
+
+  describe('Test function: updateSelectedNode', () => {
+    it('should set selectedNode', () => {
+      wrapper.vm.selectedNode = { id: 'test' };
+      wrapper.vm.updateSelectedNode({ id: 'test2' });
+      wrapper.vm.selectedNode = { id: 'test2' };
+    });
+  });
+
+  describe('Test function: onCreateFileEvent', () => {
+    it('should update activeFileTab, send ExpandFolder and OpenFile events on File', async () => {
+      wrapper.vm.activeFileTab = { isSelected: false, id: 'fileName' };
+      wrapper.vm.selectedNode = { id: 'nodeName' };
+
+      await wrapper.vm.onCreateFileEvent({ name: 'newFileName', isFolder: false });
+
+      expect(wrapper.vm.activeFileTab).toEqual({ isSelected: false, id: 'nodeName/newFileName' });
+      expect(FileEvent.ExpandFolderEvent.next).toBeCalled();
+      expect(FileEvent.OpenFileEvent.next).toBeCalled();
+    });
+
+    it('should update activeFileTab, send ExpandFolder events on Folder', async () => {
+      FileEvent.OpenFileEvent.next = jest.fn();
+      wrapper.vm.activeFileTab = { isSelected: false, id: 'fileName' };
+      wrapper.vm.selectedNode = { id: 'nodeName' };
+
+      await wrapper.vm.onCreateFileEvent({ name: 'nodeName', isFolder: true });
+
+      expect(FileEvent.ExpandFolderEvent.next).toBeCalled();
+      expect(FileEvent.OpenFileEvent.next).not.toBeCalled();
     });
   });
 
@@ -288,6 +359,18 @@ describe('Test component: ModelizerTextView', () => {
 
     it('should subscribe to CheckoutEvent', () => {
       expect(checkoutSubscribe).toHaveBeenCalledTimes(1);
+    });
+
+    it('should subscribe to SelectNodeEvent', () => {
+      expect(selectNodeSubscribe).toHaveBeenCalledTimes(1);
+    });
+
+    it('should subscribe to CreateFileEvent', () => {
+      expect(createFileSubscribe).toHaveBeenCalledTimes(1);
+    });
+
+    it('should subscribe to DeleteFileEvent', () => {
+      expect(deleteFileSubscribe).toHaveBeenCalledTimes(1);
     });
   });
 
@@ -308,6 +391,24 @@ describe('Test component: ModelizerTextView', () => {
       expect(checkoutUnsubscribe).toHaveBeenCalledTimes(0);
       wrapper.unmount();
       expect(checkoutUnsubscribe).toHaveBeenCalledTimes(1);
+    });
+
+    it('should unsubscribe to SelectNodeEvent', () => {
+      expect(selectNodeUnsubscribe).toHaveBeenCalledTimes(0);
+      wrapper.unmount();
+      expect(selectNodeUnsubscribe).toHaveBeenCalledTimes(1);
+    });
+
+    it('should unsubscribe to CreateFileEvent', () => {
+      expect(createFileUnsubscribe).toHaveBeenCalledTimes(0);
+      wrapper.unmount();
+      expect(createFileUnsubscribe).toHaveBeenCalledTimes(1);
+    });
+
+    it('should unsubscribe to DeleteNodeEvent', () => {
+      expect(deleteFileUnsubscribe).toHaveBeenCalledTimes(0);
+      wrapper.unmount();
+      expect(deleteFileUnsubscribe).toHaveBeenCalledTimes(1);
     });
   });
 });